Add unified hashing system
Implements a "semantic" hashing algorithm which hashes Thunks based on
the functional behavior of the code being executed. The intention is to
have a hash which has an identical value across different Julia sessions
for tasks which compute the same value. This is important for
implementing a "headless" worker-worker cluster, where there is no
coordinating head worker, and all workers can see the entire
computational program.

Hashes are computed automatically and can be queried with
`get_task_hash()` while running in a task context, or directly as
`get_task_hash(task)` for any Dagger task type. Hashes are also provided
within `Dagger.move` calls, where the input task's hash is also available.