docs: Simplify and streamline quickstart
The docs were far too wordy for what's supposed to be an introduction to
Dagger. This commit moves most of that information to separate deep-dive
pages, and sets up the quickstart page to have simple examples of how to
do common tasks. The hope is that new users will start by adapting the
simple example, and read more into the deep dive when they want to do
something fancy or learn how the example works.