DArray: Operations return DArrays, local partition
To enable MPI support in the DArray (which is best implemented as each
MPI rank holding only a single local partition), this commit splits
AbstractBlock further by multi-partition or single-partition storage
schemes, where `Blocks <: AbstractMultiBlocks`, and a future `MPIBlocks
<: AbstractSingleBlocks`. Additionally, a DArray ctor is added for when
only a single subdomain and chunk/thunk is provided.

For easier post-hoc repartitioning of DArrays, we now store the original
user-provided partitioning scheme within the DArray, and also add it as
a type parameter. This also assists future MPI integration by allowing
for operations to dispatch on the partitioning scheme.

Finally, this commit also adjusts all DArray operations to return
DArrays, so that no lazy operators like MatMul or Map are returned to
the user. While it would be nice to be able to work with these
operators directly, the array ecosystem has generally settled on
propagating arrays of the same or similar types as the inputs to many
operations. The operators themselves are still present behind a single
materializing call, so it should be possible to expose them again in the
future if optimization opportunities become feasible.